requirements.txt
exif~=0.11.0
Pillow~=7.0.0
numpy~=1.17.4
pandas~=1.0.4
scikit-image~=0.17.2
setuptools~=45.2.0
opencv-python~=4.2.0.34